Fix migration.

+++ database/migrations/2019_11_10_173810_switch_https_and_active_to_int.php
@@ -12,12 +12,11 @@ class SwitchHttpsAndActiveToInt extends Migration
      * @return void
      */
     public function up()
-    {
-        Schema::table('sites', function (Blueprint $table) {
-            $table->smallInteger('active')
+    {        Schema::table('sites', function (Blueprint $table) {
+            $table->integer('active')
                 ->default(1)
                 ->change();
-            $table->smallInteger('https')
+            $table->integer('https')
                 ->default(0)
                 ->change();
         });
